requirements.txt
flask 
line-bot-sdk 
BeautifulSoup4
requests
